Added possibility to configure the starting units configuration. There is a new section in the configuration files (netpanzer.ini or netpanzer-dedicated.ini)
[spawnconfig]
archer=0
bear=0
bobcat=0
drake=0
manta=2
panther1=2
spanzer=2
stinger=1
titan=2
wolf=0

If the sections doesn't exist it will be created with the default configuration (the same as the normal one). If the configured number of starting units is bigger than the max units per player, it will cut it down to the max units permited per player.