- removed lua 5.1.2 (very old) and added lua 5.1.4 (at the time of writting there is already 5.2, but it removed some functionality that we use and needs to modify code, will stick with the old one for some more time).
This time added the full source code of lua, if someone is using some IDE for editing the code, they might need to modify it to use the new lua paths. Check SConstruct file where the library is build to know what files to add.

- Modified SConstruct to be able to use the system provided lua, by default it uses the internal one. For example, to link with the system standard lua use "scons with_lua=lua". The parameter of "with_lua" will be added as a library link (-llua) so, for example "with_lua=lua5.1" will add -llua5.1.

About SCons parameters:
I wanted to use --with-lua=xxx or --with-system-lua but SCons doesn't accept the paramters this one. Now someone could say "You are wrong! You can do that in SCons!". Well doing that is deprecated and is bound to die at some random point in time.

About the Lua library change:
I was reviewing what the distros does to create netpanzer packages and found that some of them might not like the idea of using the internal libraries that are also provided by the system (in this case, first one was Debian).

This is both a risk for them and a risk for us, but just to make everybody happy, just made the modifications to allow this.

Risks for using system lua:
- if we use newer lua in our codebase than the system provided, it might not compile for them or it might do something different that will generate errors in our program. They might ask us to patch our code to work with the different lua.
- if we use older lua in our codebase, maybe it won't compile with the system provided one because the new one is not compatible. They might ask us to patch our code to use their lua.
- if the system updates the lua version, netpanzer might stop working due to incompatibilities or generate new errors that didn't happen before.

Risk for using integrated lua:
- if there is a security issue, we have to update our integrated lua ourselves, and generate an update version. Just for this point in the future, when the game is finished, we might want to use shared version of it.

Our original builds will use the integrated Lua.