In the beginning two buttons were created, Button and cButton.
For long time their heart was burning with desires of merging, but nothing was happening.

They were left alone, and ages after it was discovered that they multiplied and grow, now there were five buttons. Then the judgement day came and it was seen that their numbers generated caos.

So they got together and merged. But Button and cButton were still incompatible, so they waited for a better chance.

------

In other words:
- deleted newButton, tButton and tBitButton, functionality can be done with the standard Button
- deleted two unused views: OrderingView, SkirmishView.
- Fix: Surface::FillRoundRect was not filling correctly
- Remove some useless code in PlayerState.

The idea is to delete cButton in the near future and create some kind of "Actions" that can be called from widgets or scripts.