Fix bug when there were more than 255 messages in the chat box.
Removed the "old" chatting method.

The new:

There is an ugly button left of the chat box, if it is white you are chatting to all, if is yellow you chat with team mates or allies.

Depending on who you chat with, the characters you type will have a different color too.

Pressing "ENTER" any time will focus the chat window (doesn't change who you chat with), if the window was minimized it will be restored.

Pressing "CTRL+ENTER" will focus the chat window AND set to chat with Team mates / allies (also restore the window if minimized).